refactor: move network functions to core/network (#1703)
