+++ src/core/api/buildPayTransaction.test.ts
@@ -7,8 +7,8 @@ import {
 import {
   CDP_CREATE_PRODUCT_CHARGE,
   CDP_HYDRATE_CHARGE,
-} from '../../network/definitions/pay';
-import { sendRequest } from '../../network/request';
+} from '../network/definitions/pay';
+import { sendRequest } from '../network/request';
 /**
  * @vitest-environment node
  */
@@ -28,7 +28,7 @@ import type {
   HydrateChargeAPIParams,
 } from './types';
 
-vi.mock('../../network/request');
+vi.mock('../network/request');
 
 describe('buildPayTransaction', () => {
   afterEach(() => {
