+++ src/core/network/attestations.ts
@@ -1,7 +1,7 @@
 import { gql } from 'graphql-request';
 import { getAddress } from 'viem';
 import type { Address, Chain } from 'viem';
-import type { Attestation, EASSchemaUid } from '../identity/types';
+import type { Attestation, EASSchemaUid } from '../../identity/types';
 import { createEasGraphQLClient } from './createEasGraphQLClient';
 
 /**
