+++ src/core/network/createEasGraphQLClient.ts
@@ -1,6 +1,6 @@
 import { GraphQLClient } from 'graphql-request';
 import type { Chain } from 'viem';
-import { getChainEASGraphQLAPI } from '../identity/utils/easSupportedChains';
+import { getChainEASGraphQLAPI } from '../../identity/utils/easSupportedChains';
 
 export function createEasGraphQLClient(chain: Chain): GraphQLClient {
   const endpoint = getChainEASGraphQLAPI(chain);
