+++ src/core/network/definitions/base.ts
@@ -1,5 +1,5 @@
 import { base } from 'viem/chains';
-import type { EASChainDefinition } from '../../identity/types';
+import type { EASChainDefinition } from '../../../identity/types';
 
 export const easChainBase: EASChainDefinition = {
   id: base.id,
