+++ src/core/network/definitions/nft.ts
