+++ src/core/network/getRPCUrl.test.ts
@@ -1,6 +1,9 @@
 import { baseSepolia } from 'viem/chains';
 import { describe, expect, it } from 'vitest';
-import { getOnchainKitConfig, setOnchainKitConfig } from '../OnchainKitConfig';
+import {
+  getOnchainKitConfig,
+  setOnchainKitConfig,
+} from '../../OnchainKitConfig';
 import { getRPCUrl } from './getRPCUrl';
 
 describe('OnchainKitConfig RPC URL', () => {
