+++ src/identity/utils/getAvatar.test.tsx
@@ -1,14 +1,14 @@
 import { base, baseSepolia, mainnet, optimism } from 'viem/chains';
 import { type Mock, beforeEach, describe, expect, it, vi } from 'vitest';
-import { publicClient } from '../../network/client';
-import { getChainPublicClient } from '../../network/getChainPublicClient';
+import { publicClient } from '../../core/network/client';
+import { getChainPublicClient } from '../../core/network/getChainPublicClient';
 import { RESOLVER_ADDRESSES_BY_CHAIN_ID } from '../constants';
 import { getAvatar } from './getAvatar';
 
-vi.mock('../../network/client');
+vi.mock('../../core/network/client');
 
-vi.mock('../../network/getChainPublicClient', () => ({
-  ...vi.importActual('../../network/getChainPublicClient'),
+vi.mock('../../core/network/getChainPublicClient', () => ({
+  ...vi.importActual('../../core/network/getChainPublicClient'),
   getChainPublicClient: vi.fn(() => publicClient),
 }));
