+++ src/identity/utils/getSocials.test.ts
@@ -1,13 +1,13 @@
 import { base, mainnet } from 'viem/chains';
 import { beforeEach, describe, expect, it, vi } from 'vitest';
+import { getChainPublicClient } from '../../core/network/getChainPublicClient';
 import { isBase } from '../../isBase';
 import { isEthereum } from '../../isEthereum';
-import { getChainPublicClient } from '../../network/getChainPublicClient';
 import { getSocials } from './getSocials';
 
 vi.mock('../../isBase');
 vi.mock('../../isEthereum');
-vi.mock('../../network/getChainPublicClient');
+vi.mock('../../core/network/getChainPublicClient');
 
 describe('getSocials', () => {
   const mockClient = {
