+++ src/swap/components/SwapToast.tsx
@@ -2,9 +2,9 @@ import { useCallback } from 'react';
 import { cn, color, text } from '../../styles/theme';
 
 import { useAccount } from 'wagmi';
+import { getChainExplorer } from '../../core/network/getChainExplorer';
 import { Toast } from '../../internal/components/Toast';
 import { successSvg } from '../../internal/svg/successSvg';
-import { getChainExplorer } from '../../network/getChainExplorer';
 import type { SwapToastReact } from '../types';
 import { useSwapContext } from './SwapProvider';
