+++ src/transaction/components/TransactionButton.tsx
@@ -1,7 +1,7 @@
 import { useCallback, useMemo } from 'react';
 import { useAccount, useChainId } from 'wagmi';
+import { getChainExplorer } from '../../core/network/getChainExplorer';
 import { Spinner } from '../../internal/components/Spinner';
-import { getChainExplorer } from '../../network/getChainExplorer';
 import { border, cn, color, pressable, text } from '../../styles/theme';
 import type { TransactionButtonReact } from '../types';
 import { isSpinnerDisplayed } from '../utils/isSpinnerDisplayed';
